Fix bugs
Add face details and hair color swapping

Armor now recursively finds the valid armor race to use
(AKA race -> armorRace -> armorRace -> etc.) that the armor can support

Armor slots are modified to match the slots of the valid addons. As an example, this would allow an armor to support an addon that completely hides the head, and an addon that shows the head for different races.

Potential new bug:
Inventory may have misaccurate icon for some armors in some situations